fix: clear cache btn
