+++ packages/app/components/settings/clear-cache-btn.web.tsx
@@ -1,0 +1,3 @@
+export const ClearCacheBtn = () => {
+  return null;
+};
